adding float cast

+++ android/src/main/java/com/ysampir/flutter_scandit_plugin/ScanditView.java
@@ -122,7 +122,7 @@ public class ScanditView implements PlatformView, MethodChannel.MethodCallHandle
             _camera = Camera.getDefaultCamera();
             if (_camera != null) {
                 CameraSettings cameraSettings = BarcodeCapture.createRecommendedCameraSettings();
-                cameraSettings.setZoomFactor((double)2.0);
+                cameraSettings.setZoomFactor((float)2.0);
                 _camera.applySettings(cameraSettings);
                 _dataCaptureContext.setFrameSource(_camera);
             } else {
